code clean and organised
