+++ Frontend/src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Routes, Route, useLocation } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { ThemeProvider } from "@mui/material/styles";
 import Login from "./pages/login/Login";
-import Register from "./pages/register/Register";
+import Register from "./pages/register/registerUser/Register";
 import Notifications from "./pages/notifications/Notifications";
 import Home from "./pages/home/Home";
 import Navbar from "../src/components/navbar/Navbar";
@@ -11,7 +11,7 @@ import Announcements from "./pages/announcements/Announcements";
 import { RootState } from "./store/store";
 import Users from "./pages/users/Users";
 import AddBuilding from "./pages/addBuilding/AddBuilding";
-import AdminRegister from "./pages/register/AdminRegister";
+import AdminRegister from "./pages/register/registerManager/AdminRegister";
 import Footer from "./components/Footer/Footer";
 const currentUser = false;
 const manager = true;
