+++ Frontend/src/pages/register/registerManager/interfaces.ts
